+++ Domain/RDD.Domain/ICloneable.cs
@@ -1,7 +1,0 @@
-﻿namespace Rdd.Domain
-{
-    public interface ICloneable<T>
-    {
-        T Clone();
-    }
-}